+++ Chat/src/main/java/com/cesar/Chat/dto/UnreadMessagesDTO.java
@@ -1,10 +1,10 @@
 package com.cesar.Chat.dto;
 
-import lombok.Data;
 import java.util.UUID;
+import lombok.Data;
 
 @Data
 public class UnreadMessagesDTO {
     private UUID conversationId;
-    private Integer count;
+    private Long count;
 }
